# trie: remove redundant returns + use stacktrie where applicable