--- a/trie/proof.go
+++ b/trie/proof.go
@@ -464,115 +464,100 @@ func hasRightElement(node node, key []byte) bool {
 //
 // Except returning the error to indicate the proof is valid or not, the function will
 // also return a flag to indicate whether there exists more accounts/slots in the trie.
-func VerifyRangeProof(rootHash common.Hash, firstKey []byte, lastKey []byte, keys [][]byte, values [][]byte, proof ethdb.KeyValueReader) (ethdb.KeyValueStore, *Trie, *KeyValueNotary, bool, error) {
+func VerifyRangeProof(rootHash common.Hash, firstKey []byte, lastKey []byte, keys [][]byte, values [][]byte, proof ethdb.KeyValueReader) (ethdb.KeyValueStore, bool, error) {
 	if len(keys) != len(values) {
-		return nil, nil, nil, false, fmt.Errorf("inconsistent proof data, keys: %d, values: %d", len(keys), len(values))
+		return nil, false, fmt.Errorf("inconsistent proof data, keys: %d, values: %d", len(keys), len(values))
 	}
 	// Ensure the received batch is monotonic increasing.
 	for i := 0; i < len(keys)-1; i++ {
 		if bytes.Compare(keys[i], keys[i+1]) >= 0 {
-			return nil, nil, nil, false, errors.New("range is not monotonically increasing")
+			return nil, false, errors.New("range is not monotonically increasing")
 		}
 	}
 	// Create a key-value notary to track which items from the given proof the
 	// range prover actually needed to verify the data
-	notary := NewKeyValueNotary(proof)
+	notary := newKeyValueNotary(proof)
 
 	// Special case, there is no edge proof at all. The given range is expected
 	// to be the whole leaf-set in the trie.
 	if proof == nil {
 		var (
 			diskdb = memorydb.New()
-			triedb = NewDatabase(diskdb)
+			tr     = NewStackTrie(diskdb)
 		)
-		tr, err := New(common.Hash{}, triedb)
-		if err != nil {
-			return nil, nil, nil, false, err
-		}
 		for index, key := range keys {
 			tr.TryUpdate(key, values[index])
 		}
-		if tr.Hash() != rootHash {
-			return nil, nil, nil, false, fmt.Errorf("invalid proof, want hash %x, got %x", rootHash, tr.Hash())
+		if have, want := tr.Hash(), rootHash; have != want {
+			return nil, false, fmt.Errorf("invalid proof, want hash %x, got %x", want, have)
 		}
-		// Proof seems valid, serialize all the nodes into the database
-		if _, err := tr.Commit(nil); err != nil {
-			return nil, nil, nil, false, err
+		// Proof seems valid, serialize remaining nodes into the database
+		if _, err := tr.Commit(); err != nil {
+			return nil, false, err
 		}
-		if err := triedb.Commit(rootHash, false, nil); err != nil {
-			return nil, nil, nil, false, err
-		}
-		return diskdb, tr, notary, false, nil // No more elements
+		return diskdb, false, nil // No more elements
 	}
 	// Special case, there is a provided edge proof but zero key/value
 	// pairs, ensure there are no more accounts / slots in the trie.
 	if len(keys) == 0 {
 		root, val, err := proofToPath(rootHash, nil, firstKey, notary, true)
 		if err != nil {
-			return nil, nil, nil, false, err
+			return nil, false, err
 		}
 		if val != nil || hasRightElement(root, firstKey) {
-			return nil, nil, nil, false, errors.New("more entries available")
+			return nil, false, errors.New("more entries available")
 		}
 		// Since the entire proof is a single path, we can construct a trie and a
 		// node database directly out of the inputs, no need to generate them
 		diskdb := notary.Accessed()
-		tr := &Trie{
-			db:   NewDatabase(diskdb),
-			root: root,
-		}
-		return diskdb, tr, notary, hasRightElement(root, firstKey), nil
+		return diskdb, hasRightElement(root, firstKey), nil
 	}
 	// Special case, there is only one element and two edge keys are same.
 	// In this case, we can't construct two edge paths. So handle it here.
 	if len(keys) == 1 && bytes.Equal(firstKey, lastKey) {
 		root, val, err := proofToPath(rootHash, nil, firstKey, notary, false)
 		if err != nil {
-			return nil, nil, nil, false, err
+			return nil, false, err
 		}
 		if !bytes.Equal(firstKey, keys[0]) {
-			return nil, nil, nil, false, errors.New("correct proof but invalid key")
+			return nil, false, errors.New("correct proof but invalid key")
 		}
 		if !bytes.Equal(val, values[0]) {
-			return nil, nil, nil, false, errors.New("correct proof but invalid data")
+			return nil, false, errors.New("correct proof but invalid data")
 		}
 		// Since the entire proof is a single path, we can construct a trie and a
 		// node database directly out of the inputs, no need to generate them
 		diskdb := notary.Accessed()
-		tr := &Trie{
-			db:   NewDatabase(diskdb),
-			root: root,
-		}
-		return diskdb, tr, notary, hasRightElement(root, firstKey), nil
+		return diskdb, hasRightElement(root, firstKey), nil
 	}
 	// Ok, in all other cases, we require two edge paths available.
 	// First check the validity of edge keys.
 	if bytes.Compare(firstKey, lastKey) >= 0 {
-		return nil, nil, nil, false, errors.New("invalid edge keys")
+		return nil, false, errors.New("invalid edge keys")
 	}
 	// todo(rjl493456442) different length edge keys should be supported
 	if len(firstKey) != len(lastKey) {
-		return nil, nil, nil, false, errors.New("inconsistent edge keys")
+		return nil, false, errors.New("inconsistent edge keys")
 	}
 	// Convert the edge proofs to edge trie paths. Then we can
 	// have the same tree architecture with the original one.
 	// For the first edge proof, non-existent proof is allowed.
 	root, _, err := proofToPath(rootHash, nil, firstKey, notary, true)
 	if err != nil {
-		return nil, nil, nil, false, err
+		return nil, false, err
 	}
 	// Pass the root node here, the second path will be merged
 	// with the first one. For the last edge proof, non-existent
 	// proof is also allowed.
 	root, _, err = proofToPath(rootHash, root, lastKey, notary, true)
 	if err != nil {
-		return nil, nil, nil, false, err
+		return nil, false, err
 	}
 	// Remove all internal references. All the removed parts should
 	// be re-filled(or re-constructed) by the given leaves range.
 	empty, err := unsetInternal(root, firstKey, lastKey)
 	if err != nil {
-		return nil, nil, nil, false, err
+		return nil, false, err
 	}
 	// Rebuild the trie with the leaf stream, the shape of trie
 	// should be same with the original one.
@@ -588,16 +573,16 @@ func VerifyRangeProof(rootHash common.Hash, firstKey []byte, lastKey []byte, key
 		tr.TryUpdate(key, values[index])
 	}
 	if tr.Hash() != rootHash {
-		return nil, nil, nil, false, fmt.Errorf("invalid proof, want hash %x, got %x", rootHash, tr.Hash())
+		return nil, false, fmt.Errorf("invalid proof, want hash %x, got %x", rootHash, tr.Hash())
 	}
 	// Proof seems valid, serialize all the nodes into the database
 	if _, err := tr.Commit(nil); err != nil {
-		return nil, nil, nil, false, err
+		return nil, false, err
 	}
 	if err := triedb.Commit(rootHash, false, nil); err != nil {
-		return nil, nil, nil, false, err
+		return nil, false, err
 	}
-	return diskdb, tr, notary, hasRightElement(root, keys[len(keys)-1]), nil
+	return diskdb, hasRightElement(root, keys[len(keys)-1]), nil
 }
 
 // get returns the child of the given node. Return nil if the