@holiman 2021-04-28
> which exhibit quite a boost by the use of Stacktrie instead of the regular one.

Just to clarify: this improvement will be noticeable mainly when we're doing snap sync, and receive a small-to-medium size storage trie.